+++ plotter.hpp
@@ -42,7 +42,7 @@ public:
 
 private:
     void draw_axis(SDL2pp::Renderer& renderer);
-    void draw_point(int x, int y, SDL2pp::Renderer& renderer); // Absolute coordinates
+    void draw_point(float x, float y, SDL2pp::Renderer& renderer); // Absolute coordinates
     int compute_x_offset() const { return m_x_offset * m_zoom; }
     int compute_y_offset() const { return m_y_offset * m_zoom; }
     int to_plot_x(float x) const;
